+++ grupo1/grupo1-model/src/main/java/es/uc3m/tiw/model/dao/PromocionDAOImpl.java
@@ -58,7 +58,7 @@ public class PromocionDAOImpl implements PromocionDAO {
 	 * @see es.uc3m.tiw.daos.UsuarioDAO#recuperarUsuarioPorPK(java.lang.Integer)
 	 */
 	@Override
-	public Promocion recuperarPromocionPorPK(Long pk){
+	public Promocion recuperarPromocionPorPK(int pk){
 		return em.find(Promocion.class, pk);
 	}
 	/* (non-Javadoc)
